implemented POST /theme/class

+++ src/main/java/com/nasnav/dao/OrganizationThemeSettingsRepository.java
@@ -3,7 +3,7 @@ package com.nasnav.dao;
 import com.nasnav.persistence.OrganizationThemesSettingsEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrganizationThemeSettingsRepository extends JpaRepository<OrganizationThemesSettingsEntity, Long> {
+public interface OrganizationThemeSettingsRepository extends JpaRepository<OrganizationThemesSettingsEntity, Integer> {
 
 
 }
