+++ src/main/java/com/nasnav/dao/ThemeClassRepository.java
@@ -3,5 +3,5 @@ package com.nasnav.dao;
 import com.nasnav.persistence.ThemeClassEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ThemeClassRepository extends JpaRepository<ThemeClassEntity, Long> {
+public interface ThemeClassRepository extends JpaRepository<ThemeClassEntity, Integer> {
 }
