+++ src/main/java/com/nasnav/dao/ThemesRepository.java
@@ -3,6 +3,6 @@ package com.nasnav.dao;
 import com.nasnav.persistence.ThemeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ThemesRepository extends JpaRepository<ThemeEntity, Long> {
+public interface ThemesRepository extends JpaRepository<ThemeEntity, Integer> {
 
 }
